Write an assertion to report evaluated arguments
The macro is in tbb.core/tis. The idea is that all reusable (library)
code to implement a Clojure interpreter should go to this new namespace.

The macro has unit tests and is already integrated in the http
interpreter, so refer to the source code to see how it's used. It's very
similar to clojure.test/is, but is quiet, i.e. doesn't print anything.
Probably there is a way to re-use the `is` macro, but I couldn't figure
out how, and it was easier (and more fun) to implement my own.